chore: Zatch.OpenState ZatchUtil로 파일 분리
