+++ Zatch/Presentation/Utils/Register/Register.swift
@@ -7,14 +7,6 @@
 
 import Foundation
 
-struct Zatch{
-    @frozen
-    enum OpenState: Int{
-        case open = 1
-        case unopen = 2
-    }
-}
-
 struct Register{
     
     @frozen
@@ -30,13 +22,4 @@ struct Register{
     }
     
     typealias DateString = (year: String, month: String, date: String)
-}
-
-extension Zatch.OpenState{
-    var title: String{
-        switch self{
-        case .unopen:   return "미개봉"
-        case .open:     return "개봉"
-        }
-    }
 }
